# shop/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/apps/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Shop API
    description: The Shop service definition.
    version: 0.0.1
paths:
    /api/create/address:
        post:
            tags:
                - Shop
            operationId: Shop_CreateAddress
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateAddressReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AddressInfo'
    /api/default/address:
        put:
            tags:
                - Shop
            operationId: Shop_DefaultAddress
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AddressReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CheckResponse'
    /api/delete/address:
        post:
            tags:
                - Shop
            operationId: Shop_DeleteAddress
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AddressReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CheckResponse'
    /api/update/address:
        put:
            tags:
                - Shop
            operationId: Shop_UpdateAddress
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateAddressReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CheckResponse'
    /api/users/captcha:
        get:
            tags:
                - Shop
            operationId: Shop_Captcha
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CaptchaReply'
    /api/users/detail:
        get:
            tags:
                - Shop
            operationId: Shop_Detail
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UserDetailResponse'
    /api/users/login:
        post:
            tags:
                - Shop
            operationId: Shop_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LoginReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/RegisterReply'
    /api/users/register:
        post:
            tags:
                - Shop
            operationId: Shop_Register
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/RegisterReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/RegisterReply'
components:
    schemas:
        AddressInfo:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                name:
                    type: string
                mobile:
                    type: string
                Province:
                    type: string
                City:
                    type: string
                Districts:
                    type: string
                address:
                    type: string
                postCode:
                    type: string
                isDefault:
                    type: integer
                    format: int32
        AddressReq:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                uid:
                    type: integer
                    format: int64
        CaptchaReply:
            type: object
            properties:
                captchaId:
                    type: string
                picPath:
                    type: string
        CheckResponse:
            type: object
            properties:
                success:
                    type: boolean
        CreateAddressReq:
            type: object
            properties:
                uid:
                    type: integer
                    format: int64
                name:
                    type: string
                mobile:
                    type: string
                Province:
                    type: string
                City:
                    type: string
                Districts:
                    type: string
                address:
                    type: string
                postCode:
                    type: string
                isDefault:
                    type: integer
                    format: int32
        LoginReq:
            type: object
            properties:
                mobile:
                    type: string
                password:
                    type: string
                captcha:
                    type: string
                captchaId:
                    type: string
        RegisterReply:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                mobile:
                    type: string
                username:
                    type: string
                token:
                    type: string
                expiredAt:
                    type: integer
                    format: int64
            description: Data returned by registration and login
        RegisterReq:
            type: object
            properties:
                mobile:
                    type: string
                username:
                    type: string
                password:
                    type: string
        UpdateAddressReq:
            type: object
            properties:
                uid:
                    type: integer
                    format: int64
                name:
                    type: string
                mobile:
                    type: string
                Province:
                    type: string
                City:
                    type: string
                Districts:
                    type: string
                address:
                    type: string
                postCode:
                    type: string
                isDefault:
                    type: integer
                    format: int32
                id:
                    type: integer
                    format: int64
        UserDetailResponse:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                mobile:
                    type: string
                nickName:
                    type: string
                birthday:
                    type: integer
                    format: int64
                gender:
                    type: string
                role:
                    type: integer
                    format: int32
            description: user Detail returned
tags:
    - name: Shop